refactor(meetbot): extract a caption controller from the browser lifecycle (#645)
meetbot/engine.py owned browser startup, admission detection, caption
observation, caption-language menu driving, participant discovery, chat and
end detection side by side. Google Meet's DOM churns constantly, so every
churn meant editing one 862-line module where unrelated selector groups sat
next to each other.

Move the caption concerns into meetbot/caption_control.py (CaptionController):
the injected MutationObserver script, caption enable, and the three
caption-language selection strategies with their helpers. The shared low-level
locator helpers get one owner in meetbot/browser_control.py rather than being
duplicated. engine.py keeps the browser and session lifecycle and delegates.

engine.py 862 -> 455 lines. Pure structural move, no behaviour change: not one
selector, label, timeout or strategy order changed. Verified by comparing every
string literal in the old engine.py against the new modules — zero lost, and
the only additions are three docstrings.

meetbot/tests changed only where a test reaches a private that moved.

Closes #645
